requirements.txt
django-rest-framework==0.1.0
django==5.0.1
pymysql==1.1.0
django-cors-headers
python-dotenv==1.0.0
gunicorn==21.2.0
